remove function-scoped user and team fixtures
